+++ testapp/src/main/java/com/rakuten/tech/mobile/testapp/ui/settings/cache/Cache.kt
@@ -142,7 +142,7 @@ internal class Cache(
         private const val UNIQUE_ID_ERROR = "unique_id_error"
         private const val MESSAGING_UNIQUE_ID_ERROR = "messaging_unique_id_error"
         private const val MAUID_ERROR = "mauid_error"
-        private const val UNIVERSAL_BRIDGE_MESSAGE = "universal_bridge_error"
+        private const val UNIVERSAL_BRIDGE_MESSAGE = "universal_bridge_message"
         private const val IS_SETTING_SAVED = "is_setting_saved"
         private const val PROFILE_NAME = "profile_name"
         private const val PROFILE_PICTURE_URL = "profile_picture_url"